backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
supabase==2.0.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
PyMuPDF==1.23.8
pydantic==2.5.0
replicate==0.22.0
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
mistralai==0.0.12
# Authentication dependencies
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
# Web scraping dependencies
beautifulsoup4==4.12.2
lxml==4.9.3 